modules/findings-api/src/test/java/com/ibm/cloud/security_and_compliance_center/findings_api/v1/model/ContextTest.java
/*
 * (C) Copyright IBM Corp. 2021.
 *
 * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
 * the License. You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
 * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations under the License.
 */

package com.ibm.cloud.security_and_compliance_center.findings_api.v1.model;

import com.ibm.cloud.sdk.core.service.model.FileWithMetadata;
import com.ibm.cloud.security_and_compliance_center.findings_api.v1.model.Context;
import com.ibm.cloud.security_and_compliance_center.findings_api.v1.utils.TestUtilities;
import java.io.InputStream;
import java.util.HashMap;
import java.util.List;
import org.testng.annotations.Test;
import static org.testng.Assert.*;

/**
 * Unit test class for the Context model.
 */
public class ContextTest {
  final HashMap<String, InputStream> mockStreamMap = TestUtilities.createMockStreamMap();
  final List<FileWithMetadata> mockListFileWithMetadata = TestUtilities.creatMockListFileWithMetadata();

  @Test
  public void testContext() throws Throwable {
    Context contextModel = new Context.Builder()
      .region("testString")
      .resourceCrn("testString")
      .resourceId("testString")
      .resourceName("testString")
      .resourceType("testString")
      .serviceCrn("testString")
      .serviceName("testString")
      .environmentName("testString")
      .componentName("testString")
      .toolchainId("testString")
      .build();
    assertEquals(contextModel.region(), "testString");
    assertEquals(contextModel.resourceCrn(), "testString");
    assertEquals(contextModel.resourceId(), "testString");
    assertEquals(contextModel.resourceName(), "testString");
    assertEquals(contextModel.resourceType(), "testString");
    assertEquals(contextModel.serviceCrn(), "testString");
    assertEquals(contextModel.serviceName(), "testString");
    assertEquals(contextModel.environmentName(), "testString");
    assertEquals(contextModel.componentName(), "testString");
    assertEquals(contextModel.toolchainId(), "testString");

    String json = TestUtilities.serialize(contextModel);

    Context contextModelNew = TestUtilities.deserialize(json, Context.class);
    assertTrue(contextModelNew instanceof Context);
    assertEquals(contextModelNew.region(), "testString");
    assertEquals(contextModelNew.resourceCrn(), "testString");
    assertEquals(contextModelNew.resourceId(), "testString");
    assertEquals(contextModelNew.resourceName(), "testString");
    assertEquals(contextModelNew.resourceType(), "testString");
    assertEquals(contextModelNew.serviceCrn(), "testString");
    assertEquals(contextModelNew.serviceName(), "testString");
    assertEquals(contextModelNew.environmentName(), "testString");
    assertEquals(contextModelNew.componentName(), "testString");
    assertEquals(contextModelNew.toolchainId(), "testString");
  }
}